% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allHnew.R
\name{allHnew}
\alias{allHnew}
\title{Estimate marginal exposure-response function on new grid of points}
\usage{
allHnew(fit, Xrange = seq(0.01, 0.99, 0.01), center = NULL, alpha = 0.05)
}
\arguments{
\item{fit}{An object containing the results from \code{rffbkmr} function.}

\item{Xrange}{Vector of quantiles for new grid points.}

\item{center}{The quantile of all exposures at which the exposure–response function equals zero.}

\item{alpha}{100(1-alpha)\% posterior interval (default: 0.05).}
}
\value{
A data frame with all exposure values, posterior mean estimate, and 100(1-alpha)\% posterior interval.
}
\description{
Estimate marginal exposure-response function on new grid of points
}
\examples{
data(simudat)
fit <- rffbkmr(y = simudat$y,
X = simudat[,1:5],
Z = simudat[,6:10],
Z.eq = as.formula("~ z1 + as.factor(z2) + z3 + z4 + as.factor(z5)"),
K = 1000,
J = 10,
verbose = FALSE)
est_all <- allHnew(fit = fit)
}
