% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simudat.R
\docType{data}
\name{simudat}
\alias{simudat}
\title{Simulated Data for rffbkmr}
\format{
A list with:
\describe{
\item{y}{Numeric vector of outcomes}
\item{x}{5 continuous exposures, x1 to x5}
\item{z}{5 cocariates, z1 to z5. z2 and z5 are dummy variables and others are continuous}
}
}
\usage{
data(simudat)
}
\description{
A sample dataset with outcome, exposures, and covariates.
}
\keyword{datasets}
